+++ internal/provider/resource_company.go
@@ -10,7 +10,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/teltonika-rms/terraform-provider-teltonika-rms/internal/api"
 )
@@ -59,9 +58,6 @@ func (r *CompanyResource) Schema(ctx context.Context, req resource.SchemaRequest
 			"company_name": schema.StringAttribute{
 				Required:    true,
 				Description: "The name of the company.",
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
 			},
 			"parent_id": schema.Int64Attribute{
 				Optional:    true,